refactor(setup): split SetupWizardPage into context + per-step components (roadmap 1.3)
